requirements.txt
# FastAPI and its dependencies
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic[email]
python-multipart == 0.0.9

# Environment
python-dotenv==1.0.1
pydantic-settings==2.2.1

torch
torchvision
Pillow
transformers
